Move the cart object to other checkout objects
