+++ src/Http/Livewire/Cart.php
@@ -2,7 +2,7 @@
 
 namespace Jonassiewertsen\StatamicButik\Http\Livewire;
 
-use Jonassiewertsen\StatamicButik\Helper\Cart as ShoppingCart;
+use Jonassiewertsen\StatamicButik\Checkout\Cart as ShoppingCart;
 use Jonassiewertsen\StatamicButik\Http\Models\Product;
 use Livewire\Component;
