+++ tests/Unit/CartShippingTest.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Jonassiewertsen\StatamicButik\Tests\Unit;
+
+use Jonassiewertsen\StatamicButik\Http\Traits\MoneyTrait;
+use Jonassiewertsen\StatamicButik\Tests\TestCase;
+
+class CartShippingTest extends TestCase
+{
+    use MoneyTrait;
+
+    public function setUp(): void {
+        parent::setUp();
+    }
+}
